Add reminder banner in external UI to add licences
https://eaflood.atlassian.net/browse/WATER-4399

An issue has been raised where Licence Holders don't recognise the need to add licences to their account after a renewal and are missing their returns. There is already a feature in the service to add licences to an account so it has been agreed that a banner to highlight this is an appropriate intermediate step, before full account management is tackled.

This adds the banner.

---

Out-of-the-box GOVUK enforce that banner content does get the full with of the banner. It only gets approx 2 thirds (`max-width: 605px;` to be exact). Our designers decided that wrapping the content at 2 thirds looks 'odd' so this also involved some custom CSS to override the width of `govuk-notification-banner__content`.