Use new setup bill run journey (#2531)
https://eaflood.atlassian.net/browse/WATER-4375

> Part of a series of changes related to replacing the create bill run journey to incorporate changes for two-part tariff

To support the new two-part tariff bill run options we need to make changes to the bill run setup pages. We could have done them here but our policy is to migrate legacy functionality to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) when we have to make changes.

We need to update this legacy UI to direct a user to the new journey when they click the 'Create bill run' button. And so we can control when this happens we'll put this behind a feature toggle. That way we can be sure of the new journey before we make it available in production.