Redirect to new view after remove bill licence
https://eaflood.atlassian.net/browse/WATER-4410

In WATER-4131 we resolved an issue with the legacy service not being able to display large bill runs without erroring. We did this by replacing the existing views with new ones. When we did this though we overlooked a feature; Remove licence from a bill.

We've since [put the button back in](https://github.com/DEFRA/water-abstraction-system/pull/824) to allow users to remove a licence from a bill. But because we are still dependent on the legacy service to do the removal we are also required to use its 'processing' page.

Initial testing showed that for this scenario it was redirecting back to the old view for a bill! So, this change fixes that issue.

Whilst making this change we've also dropped the cancel button from that processing (spinner) page. It's a 'one-page-fits-all' affair hence the cancel button was always shown. However, you cannot cancel a remove bill or remove bill licence once these are in process. We never built in support for a bill run being cancelled in the middle of creation with the new SROC engines. Finally, using it with the old PRESROC often leads to 'stuck' cancelling bill runs.

So, we're taking this opportunity to remove the button and save us and the users unnecessary issues.