Fix back link on billing spinner page (#2535)
https://eaflood.atlassian.net/browse/WATER-4410

We realised recently that we'd inadvertently dropped the button that allows users to remove a bill licence from a bill. We've now put the button back in but this means the process has come under the beady eye of our QA team!

They've discovered that the spinner page often has a back link but clicking it can have varying results from 'page not found' to outright errors.

From looking at the legacy code we can see there was logic built in to make displaying the back link optional. If whatever code redirects to the spinner page doesn't include the appropriate option, the backlink will get displayed when it shouldn't.

~Or should it? A much simpler and cleaner solution that will work in all cases (the same spinner page is used for a host of reasons) is to have a fixed backlink that returns the user to the bill run they are working with.~

~This change implements that better solution (till we can drop the spinner page altogether!)~

That was going to be our solution but then we remembered. This repo includes client-side javascript that hijacks all backlinks and has them work in the same way as clicking back in the browser! 😱😭🤬 

So, plan B. We are just removing the back link altogether from this page.