Fix more missed links to legacy view licence page
https://eaflood.atlassian.net/browse/WATER-4558

> Part of the work to replace the legacy view licence page

It looks like when we did [Add feature flag for links to the water abstraction system](https://github.com/DEFRA/water-abstraction-ui/pull/2579) we missed some links that have been found in testing.

We thought we'd caught the rest in [Fix missed links to legacy view licence page](https://github.com/DEFRA/water-abstraction-ui/pull/2581) but we've found more!

Pretty telling state of affairs for the state of the code base that something as simple as this is proving so hard to nail.

Anyway, this deals with the latest missed link; from the workflow page when we display the 'Time limited' link.