Display security code in ext. change email page
https://eaflood.atlassian.net/browse/WATER-4632

We received a report that the external apps change email journey was broken. We were quickly able to confirm this was the case.

Nothing related to the fix needed to be made to this project. But such a critical feature shouldn't have been allowed to sit broken for so long.

To ensure this (till such time as we can replace the legacy version) we've [added a test](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/131) to our automated test project. The problem is the journey relies on sending an email code to the user which they then need to enter.

This change copies a solution that has been used before, where the code is displayed on the screen if the feature toggle `SHOW_VERIFICATION_CODE_FEATURE` has been enabled.

Our acceptance tests can then read this and use it to complete the journey.