Display security code in ext. change email page (#2597)
https://eaflood.atlassian.net/browse/WATER-4632

We received a report that the external app's change email journey was broken. We quickly confirmed this.

Nothing related to the fix needed to be made to this project. But such a critical feature shouldn't have been allowed to sit broken for so long.

To ensure this (until we can replace the legacy version), we've [added a test](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/131) to our automated test project. The problem is that the journey relies on sending the user an email code, which they then need to enter.

This change copies a solution that has been used before, where the code is displayed on the screen if the feature toggle `SHOW_VERIFICATION_CODE_FEATURE` has been enabled.

Our acceptance tests can then read this and use it to complete the journey.