Tag where permit is used
https://eaflood.atlassian.net/browse/WATER-4626

We want to establish what features in the legacy code depend on the permit.licence table. For example, we know in our own code, we depend on it to identify the ‘points’ for a licence.

The NALD import creates the permit licence table and document header table during the NALD import job.

The issue is that the permit.licence record is just a massive JSONB dump of all the NALD data related to the licence. It looks like the later licence import in water-abstraction-import did a much better job of breaking the same data down into constituent tables.

We’re focusing our efforts on recreating a new licence import ready for ReSP. But until we can confirm what legacy features depend on the old permit.licence table, we can’t decide whether to drop creating permit.licence records or recreate the massive JSONB dump.