Add system supplementary flag for returns
https://eaflood.atlassian.net/browse/WATER-4199

When a return is submitted or edited the user is given the chance to flag the licence for the next supplementary bill run. As part of our work to create a two-part tariff supplementary bill runs, we also need the same functionality to flag licences for our new bill run type. This PR is setting up the UI to call our endpoint in system that determines which years on the licence need to be flagged.