Amend mark for supplementary billing (#2605)
https://eaflood.atlassian.net/browse/WATER-4199
https://eaflood.atlassian.net/browse/WATER-4676

During testing for the first ticket, it was discovered that editing a return on a licence and then flagging that licence for supplementary billing, was always flagging the licence for pre-sroc supplementary billing even when the years were for sroc.

The system now handles flagging a licence for two-part tariff supplementary, so this change is allowing system to handle all the flagging for editing a return.
The postMarkLicenceForSupplementaryBilling function is being used during two flagging routes.

The first is when a return is edited. As we now handle that in system the call to the UI just passes the information for system to work out if the licence for the edited return needs to be flagged.

The second route is through the legacy licence link of recalculate bills. This functionality has also since been rebuilt in system, so whilst the code remains in case we ever need to rely on using the legacy view licence page, this route is now otherwise redundant.