Add monitoring stations feature flag for search
https://eaflood.atlassian.net/browse/WATER-4679

While adding the feature flag for the monitoring stations for the licence summary page, we realised that the
monitoring station can also be reached via the internal search tool. This could mean users are at one point in the journey being
shown the new migration monitoring station page, and at another point are shown the legacy monitoring station view.

This PR adds the feature flag functionality to the UI internal search tool to send users to the new monitoring station view.