Flag licences removed from workflow for supp billing
https://eaflood.atlassian.net/browse/WATER-4137

As part of the work to flag licences for the new two-part tariff supplementary billing, we need to be able to flag licences that are removed from workflow. If a licence is in workflow it does not get picked up for annual billing. This means if the licence is removed without any new charge information data added then we run the risk of that licence not being billed for the year. This change adds a call to our supplementary end point in the system repo, that will handle flagging the licence for two-part tariff billing if it meets the criteria to.