Flag licences removed from workflow for supp billing (#2614)
https://eaflood.atlassian.net/browse/WATER-4137

As part of the work to flag licences for the new two-part tariff supplementary billing, we need to be able to flag licences that are removed from the workflow. If a licence is in workflow it does not get picked up for annual billing. This means if the licence is removed without any new charge information added, then we risk the licence not being billed for the year. This change adds a call to our supplementary endpoint in the system repo, which will flag the licence for two-part tariff billing if it meets the criteria.