Refactor Supplementary Billing Flags
https://eaflood.atlassian.net/browse/WATER-4137
https://eaflood.atlassian.net/browse/WATER-4199
https://eaflood.atlassian.net/browse/WATER-4622
https://eaflood.atlassian.net/browse/WATER-4588

Upon completion of the above two-part tariff supplementary tickets, it was noted that there are now multiple places in the code that handle flagging for pre-sroc supplementary, sroc supplementary and now two-part tariff supplementary. The UI has its own flagging mechanisms as well as the service repo and now our new code in system. To keep things consistent, and easy to maintain it was decided that all the code for flagging a supplementary bill run be considolitated into the system repo. All other flagging mechanisms can be removed meaning there is only 1 place to update and maintain flagging. This PR is a housekeeping task to reduce the code.