Update unit tests for `ENABLE_SYSTEM_LICENCE_VIEW` feature flag (#2634)
While getting a dev environment up and running, we noticed that some UI unit tests were failing. On investigation it turned out that they fail when the `ENABLE_SYSTEM_LICENCE_VIEW` feature flag is set to `true`, which is how it's set in the `.env` files we pulled down for the dev environment. However the tests pass in CI because the flag isn't specified in `ci.yml` and it defaults to `false`.

We correct this by fixing the unit tests and explicitly setting `ENABLE_SYSTEM_LICENCE_VIEW` to `true` in `ci.yml`.