Update missed links to new view monitoring station (#2638)
https://eaflood.atlassian.net/browse/WATER-4736

In our last release, we shipped a new version of the view monitoring station page, replacing the legacy one. However, we quickly received feedback that the dates we display for tagged licences don't look right.

We've [fixed those issues](https://github.com/DEFRA/water-abstraction-system/pull/1470), but our QA team have spotted that several of the journeys started on that page still return you to the legacy page.

This change updates those links to the view monitoring station page we missed the first time to redirect to our new page based on the feature flag.