Add feature flag and link for system returns notifications (#2643)
https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy ui pages into the system repo. We need to redirect the user to the newly created pages / journey when it is ready.

This change adds a feature flag to the mangae page for the return notifications invitations link.

When the feature flag is on it will redirect the user to the new page / journey built in system