Refactor notifications setup redirect
https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy UI we are replacing the notification journey from the UI and rebuilding in system.

This change changes the notification setup target route. This is done to allow the system to set up an initial session for the journey.

Work done here - https://github.com/DEFRA/water-abstraction-system/pull/1574