Remove blipp plugin
https://github.com/DEFRA/water-abstraction-team/issues/111

We and it seems most other Defra services add this as standard to projects. But we've been (rightly) challenged on what value this adds.

On [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) we concluded it was very little, and was against our ethos of 'minimal and simple'. So [we removed it](https://github.com/DEFRA/water-abstraction-system/pull/1080).

This change removes it from this app.