Remove XML returns upload option (#2673)
https://eaflood.atlassian.net/browse/WATER-4964

To support quarterly returns, we knew we'd need to make changes to the bulk upload feature of the external site. CSV was easily understood and covered by WATER-4796 but XML caused some worry.

No one on the team has any understanding of how it works, or experience using it. It's also a pain in the 🍑 for the dev team. The package it depends on is no longer maintained. We've scrapped through a solution to allow us to upgrade the service to Node v22, but we'd have to rebuild the XML feature ( WATER-4774 ) to allow us to upgrade to the next version of Node.

All this, and there was only one known user of the feature. Or so we thought!

The billing and data team reached out to them and confirmed that, though they'd intended to use the feature, they'd never gotten around to it. So, they also use the CSV upload feature.

Therefore, we suspect no one is using this feature. To confirm this, we've devised a test. We will disable the option in the service, and then see if anyone reaches out during the April returns cycle to complain XML upload is broken for them.

If they do, we'll discuss our options. If no one complains though, we can proceed with dropping the feature and the problem XML dependencies from the service.