Always display bulk upload link on external site
https://eaflood.atlassian.net/browse/WATER-4953

Before this change, the bulk upload link in the external UI would only display if there were due returns, flagged for bulk upload, that matched the current return cycle.

Quarterly has changed the submission model from once a year to four times a year. Therefore existing functionality that switched between the summer cycle and winter cycle based on the time of year will no longer work, because the submission cycles now overlap.

We need to update the logic that determines whether to display the link, to ignore the current cycle. In short

- are there any due returns
- do they have an end date less than or equal to the current date (can't submit a return that hasn't ended)
- do any of them have the `isUpload` flag set to true

If we get a 'yes' to all three, then we need to display the link.