Always display bulk upload link on external site (#2677)
https://eaflood.atlassian.net/browse/WATER-4953

Before this change, the bulk upload link in the external UI would only display if there were due returns, flagged for bulk upload, that matched the current return cycle.

Quarterly has changed the submission model from once to four times a year. Therefore, the existing functionality that switched between the summer and winter cycles based on the time of year will no longer work because the submission cycles now overlap.

We must update the logic to determine whether to display the link or ignore the current cycle. In short

- Are there any due returns?
- Do they have an end date less than or equal to the current date? (You can't submit a return that hasn't ended.)
- Do any of them have the `isUpload` flag set to true?

We must display the link if we get a 'yes' to all three.