Fix view this return link on edit submit page
https://eaflood.atlassian.net/browse/WATER-4805

When testing something with our new submit/edit return log journey we added a submission using the legacy journey. We then spotted that it has a "View this return" link which takes you to the legacy view return log page.

To allow testing and release of the new view return log page (see WATER-4898 ) to proceed without being tied to our new submit/edit journey, we need to be able to view a return log using [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) but submit/edit it's submissions using the legacy pages.

This change fixes the link on the page to ensure it points to the right page depending on whether a new feature flag has been enabled.