Update creates CSV template module
Previously all the returns would have been for the current cycle. So, the only thing the module had to do was split them by frequency.

Now we've dropped cycle, we don't have the luxury of knowing they are all within a returns period.

So, the return logs are first grouped by due date: a short hand for grouping them by cycle.

Then, within that group they are sorted by frequency.

What gets returned is an array of CSV data sets, one for each due date and frequency combination. The rest of the changes are support generating the CSV files on this basis.