Access view billing account page via search
https://eaflood.atlassian.net/browse/WATER-4988

After completing the view billing account page in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system), it was noticed during
testing that searching for a billing account would redirect to the legacy billing account page. When a user searches for a billing account,
it should redirect to the new billing account page created.

This PR will amend searching for a billing account to redirect to the new billing account page.