Add separate Ad-Hoc notices feature flag (#2699)
https://eaflood.atlassian.net/browse/WATER-4782

Because of the need to support [quarterly return submissions](https://eaflood.atlassian.net/browse/WATER-4706) in the service from 1 April 2025, we also needed to be able to send return invitations and reminders every quarter. We've therefore replaced the legacy functionality for those journeys with our own.

We also hoped to have a new journey ready for when quarterly went live: ad-hoc notifications. However, as the needs for this journey became clearer, we'll no longer be able to meet that deadline.

Therefore, we want to make the new reminders and invitations journeys available, but keep ad-hoc hidden behind a feature flag. To do this, we need to add a new flag specifically for ad-hoc use.