Fix issue with returns csv template generation
https://eaflood.atlassian.net/browse/WATER-4796

We are updating the service to allow customers to upload returns quarterly (four times a year instead of just once).

In [Update upload CSV generation to support quarterly](https://github.com/DEFRA/water-abstraction-ui/pull/2679) we 'broke' the logic that tied the templates generated to only those in the current cycle. Instead, it is a more flexible "I have a due return; give me the CSV!"

But in testing we found a bug. Though the templates generated seemed to match the returns due, our eagle-eyed QA team spotted thta something screwy was going on with the dates.

This was one of those 'head-scratchers' as though tweaked, our code and that originally provided by @chris-barrett-ddts in [Make bulk upload work for quarterly](https://github.com/DEFRA/water-abstraction-ui/pull/2645/files) were doing the same thing.

Not surprisingly, it all came down to a variable being declared in _completely_ the wrong place! Essentially, we were generating each group correctly, but overwriting the previous with the next as we looped because of where we had declared a variable.