Enable new view notices page (#2702)
https://eaflood.atlassian.net/browse/WATER-4788

As part of our effort to replace the legacy UI, we're gradually porting pages to the new system. In [Notifications - View sent notices page](https://github.com/DEFRA/water-abstraction-system/pull/1822), we replaced the legacy 'Notices report' with a new 'Notices' page, modelled in large part on what we've done for Bill runs.

The new page is now ready to ship, so this adds a feature flag that allows us to make it visible in the next release and quickly hide it if any issues are found.