Fix redirect from charge information to billing account (#2706)
* Fix redirect from charge information to billing account

https://eaflood.atlassian.net/browse/WATER-4988

While testing fixes to the view billing account page, it was noticed there was a redirect on the charge information page that would return to the legacy billing account page. This should return to systems new billing account page.

This PR fixes the link to redirect to systems billing account page.