Fix link to billing account page from charge information
https://eaflood.atlassian.net/browse/WATER-4988

After merging in [Fix redirect from charge information to billing account](https://github.com/DEFRA/water-abstraction-ui/pull/2706), and attempting to [Update the
backlink for the billing account to system](https://github.com/DEFRA/water-abstraction-system/pull/2097), the licence id needed in the query param to redirect to the view charge
information page was undefined. This is because the link on the charge information page to the billing account page is missing the licence id as part of the query param. To fix this,
we need to add the licence id as a query argument to the link to the billing account on the view charge information page.

This PR updates the link to the billing account page from the charge information view to include the licence id query param.