Update link from customer to view billing account
https://eaflood.atlassian.net/browse/WATER-4988

We have recently migrated the legacy view billing account page to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

This has meant updating links in the legacy UI to point to the new page. Only, as usual with the legacy service finding all the links is problematic.

This change is because we have just spotted another one. This time from the **Customer** page (accessed from the licence summary contact details tab).

The **Customer** page is kind of all by itself. You can only access it from the link on the contact details tab, and it isn't licence specific (a 'customer' can be linked to multiple licences).

So, we feel obligated to allow the user to 'Go back...' to the **Customer** page from our new view billing account page. This means along with updating the link, we also need to pass through information `water-abstraction-system` can use to get the user back.