Update link from customer to view billing account (#2708)
https://eaflood.atlassian.net/browse/WATER-4988

> Depends on [Support link from customer page to billing account](https://github.com/DEFRA/water-abstraction-system/pull/2134)

We have recently migrated the legacy view billing account page to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

This has required updating links in the legacy UI to direct users to the new page. Only, as usual with the legacy service, finding all the links is problematic.

This change is because we have just spotted another one. This time from the **Customer** page (accessed from the licence summary contact details tab).

The **Customer** page is a standalone page. You can only access it from the link on the contact details tab, and it isn't licence-specific (a 'customer' can be linked to multiple licences).

Therefore, we feel obligated to allow the user to 'Go back...' to the **Customer** page from our new view billing account page. This means, along with updating the link, we also need to pass through information that the `water-abstraction-system` can use to get the user back.