Fix redirect from billing acct for chg ver setup
https://eaflood.atlassian.net/browse/WATER-5115

As part of our ongoing efforts to move away from the legacy code we recently completed work on migrating the view billing account page to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

Often, we find numerous links or redirects to a page sprinkled throughout others. One such was after a user has [completed the change address journey for a billing account](https://github.com/DEFRA/water-abstraction-ui/pull/2695). When a user clicks the 'Confirm' button at the end of the journey it will redirect them back to the view billing account page.

So, we updated that redirect to instead go to our new version when the feature flag is enabled.

The problem is the billing account address journey is _shared_ with the Charge version setup journey. In there a user has the option to setup a new billing account when adding the new charge version. If they chose too, they will be taken through the same address change journey you can access from the view billing account page.

The difference is the redirect. If you are adding a new charge version, the redirect returns you to the charge version setup journey. Or it should do. Because of how we changed the redirect, we are instead telling the service to go to a page in **water-abstraction-system** that doesn't exist.

This change fixes the redirect, so it will work for both journeys: just changing the address and setting up a new charge version.