Fix redirect from billing acct for chg ver setup (#2711)
https://eaflood.atlassian.net/browse/WATER-5115

As part of our ongoing efforts to move away from the legacy code, we recently completed work on migrating the view billing account page to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

Often, we find numerous links or redirects to a page sprinkled throughout others. One such was after a user has completed the change address journey for a billing account. When a user clicks the 'Confirm' button at the end of the journey, it will redirect them back to the view billing account page.

We [updated the redirect](https://github.com/DEFRA/water-abstraction-ui/pull/2695) to point to our new version when the feature flag is enabled.

The problem is that the billing account address journey is _shared_ with the Charge version setup journey. In there, a user has the option to set up a new billing account when adding the new charge version. If they choose to, they will be taken through the same address change journey you can access from the View Billing Account page.

The difference is the redirect. If you are adding a new charge version, the redirect returns you to the Charge Version Setup journey. Or it should do. Because of the change we made to the redirect, we are now instructing the service to navigate to a page in **water-abstraction-system** that doesn't exist.

This change fixes the redirect so that it will work for both journeys: just changing the address and setting up a new charge version.