Remove show future returns externally feature
https://eaflood.atlassian.net/browse/WATER-5006

For reasons we can't fathom, the previous team built in a feature where the external UI will show returns that are in the future to a user.

In `production` this is disabled so external users can only see return logs where the end date is less than or equal to the current date.

We've been caught out by this a few times, plus been asked to investigate why a return in the future is showing during testing, only to remember this quirk in the app.

Whilst working on WATER-5006 it happened again. 5 minutes of panic trying to figure out why something from the future is showing only to find end date is dynamically applied to the filter because of this toggle.

Well, enough is enough!