Remove show future returns externally feature (#2712)
https://eaflood.atlassian.net/browse/WATER-5006

For reasons we cannot fathom, the previous team built a feature that allows the external UI to display returns that are in the future to a user.

In `production`, this is disabled so external users can only see return logs where the end date is less than or equal to the current date.

We've been caught out by this a few times, and we've also been asked to investigate why a return in the future is showing during testing, only to recall this quirk in the app.

While working on WATER-5006 , it happened again. Five minutes of panic ensued as we tried to figure out why something from the future was showing, only to discover that the end date was dynamically applied to the filter due to this toggle.

Well, enough is enough!