Fix failing unit tests
Spotted that when running the unit tests locally the `manage-nav.test.js` would always fail. No doubt this is due to changes we've been making as we migrate legacy journeys that start from the Manage page to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

But also looks like we didn't stub the config properly. So, it might be fine in CI or on a specific dev's machine. But when run another machine it fails because the config is set differently.

This change should ensure they run as expected in all environments.