Fix failing unit tests (#2713)
Spotted that when running the unit tests locally, the `manage-nav.test.js` would always fail. No doubt this is due to changes we've been making as we migrate legacy journeys that start from the Manage page to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

Initially, we thought it was because we hadn't stubbed the config properly. However, we then noticed that the `manage-nav.js` module was requiring the same configuration twice, using both standard and destructuring methods.

This change sorts the issue and gets the tests running again in all environments.