Use new link and tag for returns in search results
https://eaflood.atlassian.net/browse/WATER-5006

There are two key changes we make here.

The first is the link. We've recently migrated the legacy view return log page to a new one in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system), which means we need the search results to use this new link when listing returns.

The second is the status badge.

As part of WRLS taking over returns management from NALD is a change in how return logs are generated. Now they'll be created for the current cycle when a new return version is added, rather than only for the previous return cycle. This is to help us support quarterly returns.

This means we now have returns logs which are 'not yet due' to be submitted. We created a new badge in WATER-4954 to display when listing or viewing these return logs in **water-abstraction-system**. So, we need to make the same change to the UI for the search results.