Fix ext view returns showing 'not due yet' returns
https://eaflood.atlassian.net/browse/WATER-5128

When testing WATER-5117 our QA team were looking at a water company licence (03/28/38/0018) that had its quarterly return logs all setup for the current year. In this case they have three return requirements, so three return logs have been generated per quarter. That means 12 in total for the period 2025-04-01 to 2026-03-31.

Logging in as an external user on 30 June 2025, they found they could see the first batch of quarterly return logs (1 April to June 30) all with a status of 'NOT DUE YET'. This was not expected.

The reason is a faulty assumption in the logic that determines whether to show a return log. If a return log's end date is less than _or equal to_ the current date it will show it. It should be just less than the current date.

In our example, the return log can be submitted from 1 July not 30 June when it ends.

This change fixes that logic.